First demo version of micro bit
This version supports reading a value from pin 0 on button B press.
When button A is down, the display is cleared.
When button A then goes up the microbit goes into power off mode
In power off mode it pulls 200uA.   The amount of current drops as
time goes on, for some reason.